Move network.py Mock classes to sandbox.py
